+++ crates/cargo-contract/templates/tools/generate-metadata/main.rs
@@ -3,7 +3,7 @@ extern crate contract;
 extern "Rust" {
     // Note: The ink! metdata codegen generates an implementation for this function,
     // which is what we end up linking to here.
-    fn __ink_generate_metadata() -> ink_metadata::InkProject;
+    fn __ink_generate_metadata() -> ink::metadata::InkProject;
 }
 
 fn main() -> Result<(), std::io::Error> {